backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary      # allerede brugt af db.py
python-dotenv
